pyproject.toml
[tool.poetry]
name = "vioratalk"
version = "0.4.0"
description = "AI Voice Assistant with Emotional Expression"
authors = ["MizuiroDeep <36126374+MizuiroDeep@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "vioratalk", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11.9"
apscheduler = "^3.11.0"
pyyaml = "^6.0.2"
typing-extensions = "^4.14.1"
numpy = "^2.3.2"
faster-whisper = "^1.2.0"
pyttsx3 = "^2.99"
aiohttp = "^3.12.15"
tenacity = "^9.1.2"
sounddevice = "^0.5.2"
keyboard = "^0.13.5"
google-genai = "^1.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"
pre-commit = "^3.5.0"
pytest-asyncio = "0.21.0"
pytest-cov = "^6.2.1"
ruff = "^0.12.9"
aioresponses = "^0.7.8"
pywin32 = "^311"
psutil = "^7.0.0"
pytest-xdist = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=vioratalk",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-branch",
    "-v",
]
asyncio_mode = "auto"
markers = [
    "unit: 単体テスト",
    "integration: 統合テスト",
    "e2e: エンドツーエンドテスト",
    "slow: 実行に時間がかかるテスト",
    "phase: 特定のPhaseでのテスト",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
# Phase 1の目標カバレッジ: 60%
fail_under = 60